Troubleshooting for Path Value and Time Spent
Added a flag, named isInMiddleOfCourse, which turns true whenever the
user is in between the graph, and not getting the recommendation from
the starting vertex. Added the code inside ObjectiveFunction.pathValue
(.), since this was the function called by takeBestChoice (.) is calling
this function before doing the comparison. Hence, adjusting the
iterations would give an optimal answer for any particular learner.
Using this way, both the targets are achieved such that the time is
considered only for the upcoming vertices, and also the same vertex is
not considered more than once while computing the path value. So, while
the comparison only that portion of the path is taken up that the user
is recommended to take up in future, however, for the display purpose,
path is shown from his current vertex every time.
Unicast me in case of any doubts ...